requirements.txt
APScheduler==2.1.2
w1thermsensor==1.0.4
RPi.GPIO==0.6.3
smbus==1.1
SimpleCV==1.3
-e git+https://github.com/mapler/DHT11_Python@feature/package#egg=dht11_python